% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{Collection}
\alias{Collection}
\title{Collection}
\description{
Creates an Collection class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Collection-new}{\code{Collection$new()}}
\item \href{#method-Collection-aliases}{\code{Collection$aliases()}}
\item \href{#method-Collection-create}{\code{Collection$create()}}
\item \href{#method-Collection-delete}{\code{Collection$delete()}}
\item \href{#method-Collection-exists}{\code{Collection$exists()}}
\item \href{#method-Collection-info}{\code{Collection$info()}}
\item \href{#method-Collection-list}{\code{Collection$list()}}
\item \href{#method-Collection-clone}{\code{Collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-new"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$new(req)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-aliases"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-aliases}{}}}
\subsection{Method \code{aliases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$aliases(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-create"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$create(
  name,
  ...,
  shard_number = NULL,
  sharding_method = NULL,
  replication_factor = NULL,
  write_consistency_factor = NULL,
  read_fan_out_factor = NULL,
  on_disk_payload = NULL,
  sparse_vectors = NULL,
  hnsw_config = NULL,
  optimizer_config = NULL,
  wal_config = NULL,
  quantization_config = NULL,
  init_from = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$delete(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-exists"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$exists(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-info"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$info(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-list"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
