% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_manager.R
\docType{class}
\name{CollectionManager}
\alias{CollectionManager}
\title{CollectionManager Class}
\format{
An R6 class object.
}
\description{
This class manages collections in the Qdrant database. It provides methods
to create, retrieve details, list, check existence, and delete collections.
}
\section{Methods}{

\describe{
  \item{\code{initialize(client)}}{Creates a new CollectionManager object with the given Qdrant client.}
  \item{\code{create(collection_name, vector_size)}}{Creates a new collection with the specified name and vector size.}
  \item{\code{detail(collection_name)}}{Retrieves the details of a specific collection by its name.}
  \item{\code{list()}}{Lists all available collections.}
  \item{\code{exists(collection_name)}}{Checks if a specific collection exists.}
  \item{\code{delete(collection_name)}}{Deletes the specified collection by name.}
}
}

\examples{
# Assuming `client` is an instance of QdrantClient
manager <- CollectionManager$new(client)

# Create a collection
manager$create("example_collection", 128)

# Get details of a collection
details <- manager$detail("example_collection")

# List all collections
collections <- manager$list()

# Check if a collection exists
exists <- manager$exists("example_collection")

# Delete a collection
manager$delete("example_collection")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{An instance of the \code{QdrantClient} class used to make API requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CollectionManager-new}{\code{CollectionManager$new()}}
\item \href{#method-CollectionManager-create}{\code{CollectionManager$create()}}
\item \href{#method-CollectionManager-detail}{\code{CollectionManager$detail()}}
\item \href{#method-CollectionManager-list}{\code{CollectionManager$list()}}
\item \href{#method-CollectionManager-exists}{\code{CollectionManager$exists()}}
\item \href{#method-CollectionManager-delete}{\code{CollectionManager$delete()}}
\item \href{#method-CollectionManager-clone}{\code{CollectionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$new(client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-create"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$create(collection_name, vector_size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-detail"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-detail}{}}}
\subsection{Method \code{detail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$detail(collection_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-list"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-exists"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-exists}{}}}
\subsection{Method \code{exists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$exists(collection_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-delete"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$delete(collection_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CollectionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CollectionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CollectionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
