% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{geo_bounding_box}
\alias{geo_bounding_box}
\title{Filter by geographic bounding box}
\usage{
geo_bounding_box(
  key,
  top_left_lon,
  top_left_lat,
  bottom_right_lon,
  bottom_right_lat
)
}
\arguments{
\item{key}{Field name (must be a geo field).}

\item{top_left_lon}{Longitude of the top-left corner.}

\item{top_left_lat}{Latitude of the top-left corner.}

\item{bottom_right_lon}{Longitude of the bottom-right corner.}

\item{bottom_right_lat}{Latitude of the bottom-right corner.}
}
\value{
A condition for use inside \code{must()}, \code{should()}, or \code{must_not()}.
}
\description{
Filter by geographic bounding box
}
\examples{
must(geo_bounding_box("location",
  top_left_lon = -74.0, top_left_lat = 40.8,
  bottom_right_lon = -73.9, bottom_right_lat = 40.7))
}
