% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{geo_radius}
\alias{geo_radius}
\title{Filter by geographic radius}
\usage{
geo_radius(key, lon, lat, radius)
}
\arguments{
\item{key}{Field name (must be a geo field).}

\item{lon}{Longitude of the center point.}

\item{lat}{Latitude of the center point.}

\item{radius}{Radius in meters.}
}
\value{
A condition for use inside \code{must()}, \code{should()}, or \code{must_not()}.
}
\description{
Filter by geographic radius
}
\examples{
must(geo_radius("location", lon = -73.99, lat = 40.73, radius = 500))
}
