% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{is_empty_field}
\alias{is_empty_field}
\title{Match points where a payload field is empty}
\usage{
is_empty_field(key)
}
\arguments{
\item{key}{Field name.}
}
\value{
A condition for use inside \code{must()}, \code{should()}, or \code{must_not()}.
}
\description{
Match points where a payload field is empty
}
\examples{
must(is_empty_field("tags"))
}
