% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{is_null_field}
\alias{is_null_field}
\title{Match points where a payload field is NULL or missing}
\usage{
is_null_field(key)
}
\arguments{
\item{key}{Field name.}
}
\value{
A condition for use inside \code{must()}, \code{should()}, or \code{must_not()}.
}
\description{
Match points where a payload field is NULL or missing
}
\examples{
must(is_null_field("description"))
}
