% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{must_not}
\alias{must_not}
\title{None of the conditions may match}
\usage{
must_not(...)
}
\arguments{
\item{...}{One or more conditions.}
}
\value{
A clause object for use in \code{qdrant_filter()}.
}
\description{
None of the conditions may match
}
\examples{
must_not(be("color", "black"), be("color", "brown"))
}
