% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{qdrant_client}
\alias{qdrant_client}
\title{Main qdrant_client module}
\format{
qdrant_client module
}
\usage{
qdrant_client
}
\description{
Interface to main qdrant_client  module. Provides access to top level classes
and functions as well as sub-modules
}
\keyword{datasets}
