% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{qdrant_filter}
\alias{qdrant_filter}
\title{Build a Qdrant filter from clause functions}
\usage{
qdrant_filter(...)
}
\arguments{
\item{...}{One or more clause objects created by \code{must()}, \code{should()}, or \code{must_not()}.}
}
\value{
A named list suitable for the \code{filter} parameter.
}
\description{
Build a Qdrant filter from clause functions
}
\examples{
qdrant_filter(
  must(be("color", "red")),
  must_not(be("color", "black"))
)
}
