% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{query_nearest}
\alias{query_nearest}
\title{Build a nearest-neighbour query}
\usage{
query_nearest(vector)
}
\arguments{
\item{vector}{Numeric vector to search by.}
}
\value{
A query list for \code{query_points}.
}
\description{
Build a nearest-neighbour query
}
\examples{
query_nearest(c(0.1, 0.2, 0.3))
}
