% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_manager.R
\docType{class}
\name{ServiceManager}
\alias{ServiceManager}
\title{ServiceManager Class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods to interact with the Qdrant API.
}
\description{
This R6 class provides methods to interact with the Qdrant API, including
telemetry collection, Prometheus metrics, Kubernetes health checks,
and write protection settings.
}
\section{Methods}{


\describe{
  \item{\code{new(client)}}{Constructor method. Initializes the `ServiceManager` class with a Qdrant client.}
  \item{\code{collect_telemetry_data()}}{Collects telemetry data from the Qdrant instance.
  Sends a GET request to `/telemetry/` endpoint.}
  \item{\code{retrieve_instance_details()}}{Retrieves instance details from the Qdrant base URL.
  Sends a GET request to the base URL.}
  \item{\code{collect_prometheus_metrics(anonymize = TRUE)}}{Collects Prometheus metrics from the Qdrant instance.
  Sends a GET request to `/metrics/` with an optional `anonymize` query parameter (default: TRUE).}
  \item{\code{check_write_protection()}}{Checks the current write protection status of the Qdrant instance.
  Sends a GET request to `/locks/`.}
  \item{\code{set_write_protection(write = TRUE)}}{Sets the write protection status of the Qdrant instance.
  Sends a POST request to `/locks/` with a body parameter `write`.}
  \item{\code{kubernetes_health_check()}}{Performs a Kubernetes health check on the Qdrant instance.
  Sends a GET request to `/healthz/`.}
  \item{\code{kubernetes_liveness_probe()}}{Performs a Kubernetes liveness probe on the Qdrant instance.
  Sends a GET request to `/livez/`.}
  \item{\code{kubernetes_readiness_probe()}}{Performs a Kubernetes readiness probe on the Qdrant instance.
  Sends a GET request to `/readyz/`.}
}
}

\examples{
# Assuming `client` is an initialized Qdrant client object
service_manager <- ServiceManager$new(client)
service_manager$collect_telemetry_data()
service_manager$retrieve_instance_details()
service_manager$collect_prometheus_metrics(anonymize = TRUE)
service_manager$check_write_protection()
service_manager$set_write_protection(write = FALSE)
service_manager$kubernetes_health_check()
service_manager$kubernetes_liveness_probe()
service_manager$kubernetes_readiness_probe()
}
\keyword{data}
\keyword{kubernetes}
\keyword{locks}
\keyword{prometheus}
\keyword{telemetry}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceManager-new}{\code{ServiceManager$new()}}
\item \href{#method-ServiceManager-collect_telemetry_data}{\code{ServiceManager$collect_telemetry_data()}}
\item \href{#method-ServiceManager-retrieve_instance_details}{\code{ServiceManager$retrieve_instance_details()}}
\item \href{#method-ServiceManager-collect_prometheus_metrics}{\code{ServiceManager$collect_prometheus_metrics()}}
\item \href{#method-ServiceManager-check_write_protection}{\code{ServiceManager$check_write_protection()}}
\item \href{#method-ServiceManager-set_write_protection}{\code{ServiceManager$set_write_protection()}}
\item \href{#method-ServiceManager-kubernetes_health_check}{\code{ServiceManager$kubernetes_health_check()}}
\item \href{#method-ServiceManager-kubernetes_liveness_probe}{\code{ServiceManager$kubernetes_liveness_probe()}}
\item \href{#method-ServiceManager-kubernetes_readiness_probe}{\code{ServiceManager$kubernetes_readiness_probe()}}
\item \href{#method-ServiceManager-clone}{\code{ServiceManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$new(client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-collect_telemetry_data"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-collect_telemetry_data}{}}}
\subsection{Method \code{collect_telemetry_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$collect_telemetry_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-retrieve_instance_details"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-retrieve_instance_details}{}}}
\subsection{Method \code{retrieve_instance_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$retrieve_instance_details()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-collect_prometheus_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-collect_prometheus_metrics}{}}}
\subsection{Method \code{collect_prometheus_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$collect_prometheus_metrics(anonymize = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-check_write_protection"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-check_write_protection}{}}}
\subsection{Method \code{check_write_protection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$check_write_protection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-set_write_protection"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-set_write_protection}{}}}
\subsection{Method \code{set_write_protection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$set_write_protection(write = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-kubernetes_health_check"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-kubernetes_health_check}{}}}
\subsection{Method \code{kubernetes_health_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$kubernetes_health_check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-kubernetes_liveness_probe"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-kubernetes_liveness_probe}{}}}
\subsection{Method \code{kubernetes_liveness_probe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$kubernetes_liveness_probe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-kubernetes_readiness_probe"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-kubernetes_readiness_probe}{}}}
\subsection{Method \code{kubernetes_readiness_probe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$kubernetes_readiness_probe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
