% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{should}
\alias{should}
\title{At least one condition must match}
\usage{
should(...)
}
\arguments{
\item{...}{One or more conditions.}
}
\value{
A clause object for use in \code{qdrant_filter()}.
}
\description{
At least one condition must match
}
\examples{
should(be("color", "red"), be("color", "blue"))
}
