% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collections.R
\name{Collections}
\alias{Collections}
\title{Collections}
\description{
This class provides methods to manage collections in Qdrant.
It allows for creating, retrieving, listing, checking existence, and deleting collections.
}
\examples{

## ------------------------------------------------
## Method `Collections$create_collection`
## ------------------------------------------------

\dontrun{
  client <- QdrantClient$new()
  collections <- Collections$new(client)
  collections$create_collection("my_collection", 128)
}

## ------------------------------------------------
## Method `Collections$get_collection_details`
## ------------------------------------------------

\dontrun{
  collections$get_collection_details("my_collection")
}

## ------------------------------------------------
## Method `Collections$list_all_collections`
## ------------------------------------------------

\dontrun{
  collections$list_all_collections()
}

## ------------------------------------------------
## Method `Collections$check_collection_existence`
## ------------------------------------------------

\dontrun{
  collections$check_collection_existence("my_collection")
}

## ------------------------------------------------
## Method `Collections$delete_collection`
## ------------------------------------------------

\dontrun{
  collections$delete_collection("my_collection")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{A `QdrantClient` object to handle the connection and API requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Collections-new}{\code{Collections$new()}}
\item \href{#method-Collections-create_collection}{\code{Collections$create_collection()}}
\item \href{#method-Collections-get_collection_details}{\code{Collections$get_collection_details()}}
\item \href{#method-Collections-list_all_collections}{\code{Collections$list_all_collections()}}
\item \href{#method-Collections-check_collection_existence}{\code{Collections$check_collection_existence()}}
\item \href{#method-Collections-delete_collection}{\code{Collections$delete_collection()}}
\item \href{#method-Collections-clone}{\code{Collections$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-new"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Collections object with a Qdrant client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{An instance of the `QdrantClient` class to interact with the Qdrant API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-create_collection"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-create_collection}{}}}
\subsection{Method \code{create_collection()}}{
Create a new collection in Qdrant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$create_collection(collection_name, vector_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_name}}{A character string representing the name of the collection to create.}

\item{\code{vector_size}}{An integer specifying the size of vectors stored in the collection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the API, indicating whether the collection was created.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  client <- QdrantClient$new()
  collections <- Collections$new(client)
  collections$create_collection("my_collection", 128)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-get_collection_details"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-get_collection_details}{}}}
\subsection{Method \code{get_collection_details()}}{
Get details of a specific collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$get_collection_details(collection_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_name}}{A character string representing the name of the collection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the details of the specified collection.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  collections$get_collection_details("my_collection")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-list_all_collections"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-list_all_collections}{}}}
\subsection{Method \code{list_all_collections()}}{
List all collections in the Qdrant instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$list_all_collections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of collections stored in Qdrant.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  collections$list_all_collections()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-check_collection_existence"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-check_collection_existence}{}}}
\subsection{Method \code{check_collection_existence()}}{
Check if a specific collection exists in Qdrant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$check_collection_existence(collection_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_name}}{A character string representing the name of the collection to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical value indicating whether the collection exists (TRUE/FALSE).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  collections$check_collection_existence("my_collection")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-delete_collection"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-delete_collection}{}}}
\subsection{Method \code{delete_collection()}}{
Delete a collection from Qdrant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$delete_collection(collection_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_name}}{A character string representing the name of the collection to delete.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the API indicating whether the collection was successfully deleted.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  collections$delete_collection("my_collection")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collections-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Collections-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collections$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
